+++ src/pyhilbert/utils.py
@@ -12,6 +12,7 @@ from typing import (
 )
 import torch
 import numpy as np
+from .precision import global_float_dtype, global_np_float_dtype
 
 
 class FrozenDict(Mapping):
@@ -163,9 +164,9 @@ def generate_k_path(
     pts_np = {}
     for k, v in points.items():
         if isinstance(v, torch.Tensor):
-            pts_np[k] = v.detach().cpu().numpy().astype(float)
+            pts_np[k] = v.detach().cpu().numpy().astype(global_np_float_dtype)
         else:
-            pts_np[k] = np.array(v, dtype=float)
+            pts_np[k] = np.array(v, dtype=global_np_float_dtype)
 
     for i in range(len(path_labels) - 1):
         start_label = path_labels[i]
@@ -190,16 +191,16 @@ def generate_k_path(
         next_idx = len(k_vecs_list) - 1
         node_indices.append(next_idx)
 
-    k_vecs = torch.tensor(np.array(k_vecs_list), dtype=torch.float64)
+    k_vecs = torch.tensor(np.array(k_vecs_list), dtype=global_float_dtype)
 
     # Recalculate distances precisely from the vectors
     if len(k_vecs) > 0:
         diffs = torch.norm(k_vecs[1:] - k_vecs[:-1], dim=1)
         k_dist = torch.cat(
-            [torch.tensor([0.0], dtype=torch.float64), torch.cumsum(diffs, dim=0)]
+            [torch.tensor([0.0], dtype=global_float_dtype), torch.cumsum(diffs, dim=0)]
         )
     else:
-        k_dist = torch.tensor([], dtype=torch.float64)
+        k_dist = torch.tensor([], dtype=global_float_dtype)
 
     return k_vecs, k_dist, node_indices
